mudata2: Interchange Tools for Multi-Parameter Spatiotemporal Data

Formatting and structuring multi-parameter spatiotemporal data is often a time-consuming task. This package offers functions and data structures designed to easily organize and visualize these data for applications in geology, paleolimnology, dendrochronology, and paleoclimate. See Dunnington and Spooner (2018) <doi:10.1139/facets-2017-0026>.

See also