Combine mudata objects

This implmentation of rbind combines component tables using bind_rows and distinct. When combined object use different datasets, or when subsets of the same object are recombined, this function works well. When this is not the case, it may be necessary to modify the tables such that when they are passed to bind_rows and distinct, no duplicate information exists. This should be picked up by validate_mudata.

# S3 method for mudata
rbind(..., validate = TRUE)

Arguments

...
mudata objects to combine

validate
Flag to validate the final object using validate_mudata.

Value

A mudata object

Examples

rbind(
  kentvillegreenwood %>%
    select_params(maxtemp) %>%
    select_locations(starts_with("KENT")),
  kentvillegreenwood %>%
    select_params(mintemp) %>%
    select_locations(starts_with("GREEN"))
)

#> A mudata object aligned along "date"
#>   distinct_datasets():  "ecclimate"
#>   distinct_locations(): "GREENWOOD A", "KENTVILLE CDA CS"
#>   distinct_params():    "maxtemp", "mintemp"
#>   src_tbls():           "data", "locations" ... and 3 more
#> 
#> tbl_data() %>% head():
#> # A tibble: 6 x 6
#>   dataset   location         param   date       value flags
#>   <chr>     <chr>            <chr>   <date>     <dbl> <chr>
#> 1 ecclimate KENTVILLE CDA CS maxtemp 1999-07-01  28.5 <NA> 
#> 2 ecclimate KENTVILLE CDA CS maxtemp 1999-07-02  30.7 <NA> 
#> 3 ecclimate KENTVILLE CDA CS maxtemp 1999-07-03  26.4 <NA> 
#> 4 ecclimate KENTVILLE CDA CS maxtemp 1999-07-04  28.6 <NA> 
#> 5 ecclimate KENTVILLE CDA CS maxtemp 1999-07-05  26   <NA> 
#> 6 ecclimate KENTVILLE CDA CS maxtemp 1999-07-06  25.3 <NA>