A (Mostly) Universal Data Format for Multi-Parameter, Spatiotemporal Data

The 'mudata' package for R is a set of tools to create, manipulate, and visualize multi-parameter, spatiotemporal data. Data of this type includes all data where multiple parameters (e.g. wind speed, precipitation, temperature) are measured along common axes (e.g. time, depth) at discrete locations (e.g. climate stations). These data include long-term climate data collected from climate stations, paleolimnological data, ice core data, long-term water quality monitoring data, and ocean core data among many others. Data of this type is often voluminous and difficult to organize/document given its multidimensional nature. The (mostly) universal data (mudata) format is an attempt to organize these data in a common way to facilitate their documentation and comparison.

Details

The (mostly) universal data format is a collection of five (or more) tables, one of which contains the data in a parameter-long form (see gather). The easiest way to visualize a mudata object is to inspect the sample datasets within the package (ns_climate, kentvillegreenwood, alta_lake, long_lake, and second_lake_temp).

References

Dunnington DW and Spooner IS (2018). "Using a linked table-based structure to encode self-describing multiparameter spatiotemporal data". FACETS. doi:10.1139/facets-2017-0026

See also

Examples

print(kentvillegreenwood)

#> A mudata object aligned along "date"
#>   distinct_datasets():  "ecclimate"
#>   distinct_locations(): "GREENWOOD A", "KENTVILLE CDA CS"
#>   distinct_params():    "cooldegdays", "dirofmaxgust" ... and 9 more
#>   src_tbls():           "data", "locations" ... and 3 more
#> 
#> tbl_data() %>% head():
#> # A tibble: 6 x 6
#>   dataset   location         param   date       value flags
#>   <chr>     <chr>            <chr>   <date>     <dbl> <chr>
#> 1 ecclimate KENTVILLE CDA CS maxtemp 1999-07-01  28.5 <NA> 
#> 2 ecclimate KENTVILLE CDA CS maxtemp 1999-07-02  30.7 <NA> 
#> 3 ecclimate KENTVILLE CDA CS maxtemp 1999-07-03  26.4 <NA> 
#> 4 ecclimate KENTVILLE CDA CS maxtemp 1999-07-04  28.6 <NA> 
#> 5 ecclimate KENTVILLE CDA CS maxtemp 1999-07-05  26   <NA> 
#> 6 ecclimate KENTVILLE CDA CS maxtemp 1999-07-06  25.3 <NA>